pyproject.toml
[project]
name = "lpk-price-log-ver2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.7",
    "gspread>=6.2.1",
    "httpx>=0.28.1",
    "pydantic>=2.11.4",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.3",
    "tenacity>=9.1.4",
]
